refactor(ui): remove radar chart and update task card component
